# server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Uptime Kuma API
  description: |
    REST API for Uptime Kuma — a self-hosted monitoring tool.

    ## Authentication
    All `/api/v1/*` endpoints require an API key via the `Authorization` header:
    ```
    Authorization: Bearer uk{id}_{key}
    ```
    Or via query parameter: `?apikey=uk{id}_{key}`

    Create API keys in **Settings → API Keys** and assign scopes.

    ## Roles
    - **admin** — Full access to all resources including users and settings.
    - **read-only** — Read-only access to monitors, notifications, status pages, maintenance, tags, uptime, and docker hosts.

    ## Scopes
    API keys have scopes that restrict which endpoints they can access.
    Admin-only scopes (`users:read`, `users:write`, `settings:read`, `settings:write`) are
    automatically stripped from keys owned by non-admin users.

    ## Public Endpoints
    Entry page, push heartbeats, badges, and public status pages do **not** require authentication.
  version: 2.1.3
  license:
    name: MIT
    url: https://github.com/louislam/uptime-kuma/blob/master/LICENSE
servers:
  - url: /
    description: Same-origin (relative to your Uptime Kuma instance)
...
paths:
...
  /api/status-page/{slug}:
    get:
      tags: [Status Page (Public)]
      summary: Get status page data
      description: Config, incidents, and monitor list for a published status page.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status page JSON
          content:
            application/json:
              schema:
                type: object
                description: Contains config, incidentList, monitorList, etc.
        "404":
          description: Status page not found or not published
  /api/status-page/heartbeat/{slug}:
    get:
      tags: [Status Page (Public)]
      summary: Status page heartbeat/polling data
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Heartbeat and uptime data per monitor
          content:
            application/json:
              schema:
                type: object
  /api/status-page/{slug}/incident-history:
    get:
      tags: [Status Page (Public)]
      summary: Incident history
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
      responses:
        "200":
          description: List of incidents with optional next cursor
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  incidents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "API key as Bearer token: Authorization: Bearer uk{id}_{key}"
...
  schemas:
...
    Incident:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        style:
          type: string
          enum: [info, warning, danger, primary, light, dark]
        pin:
          type: boolean
        active:
          type: boolean
        createdDate:
          type: string
          format: date-time
        lastUpdatedDate:
          type: string
          format: date-time
...